refactor the python project structure
