+++ .gitignore
@@ -1,3 +1,6 @@
 **/node_modules
 coverage
 .nyc_output
+**__pycache__
+.vscode
+env
