+++ Makefile
@@ -1,0 +1,5 @@
+init:
+	pip install -r requirements.txt
+
+test:
+	nosetests -v tests
