+++ requirements.txt
@@ -1,0 +1,3 @@
+requests
+bs4
+nose
